update naming packages & files
